services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-control-plane-api"
version = "0.3.0"
description = "FastAPI control plane service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "redis>=5.1.1",
  "pydantic-settings>=2.6.1",
  "qiskit>=1.2.0",
  "qiskit-ibm-runtime>=0.36.1",
  "prometheus-client>=0.21.0",
  "opentelemetry-sdk>=1.30.0",
  "opentelemetry-exporter-otlp>=1.30.0",
  "opentelemetry-instrumentation-fastapi>=0.51b0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "mypy>=1.11.0",
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.24.0",
  "fakeredis>=2.25.0",
  "aiosqlite>=0.20.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
ignore_missing_imports = true
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"
markers = ["integration: integration tests requiring postgres/redis"]
